refactor(foundation): move IntelligentBehaviorConfig + BehaviorRules family (A8)
Phase 6 / Sub-phase A8: Major expansion of foundation. Now that ConsistencyRule
and StateMachine are in foundation (A3), BehaviorRules became extractable —
which cascades to BehaviorModelConfig, IntelligentBehaviorConfig, and the
related config types.

Moved to mockforge-foundation::intelligent_behavior:
- types::BehaviorRules
- config::IntelligentBehaviorConfig
- config::BehaviorModelConfig
- config::VectorStoreConfig
- config::PerformanceConfig
- config::PersonasConfig

Core's intelligent_behavior/config.rs and types.rs are now thin re-export shims.

Migrated consumers to foundation imports:
- mockforge-chaos: api.rs IntelligentBehaviorConfig (still blocked by MockAI)
- mockforge-http: handlers/ai_studio.rs
- mockforge-ui: handlers/{ai_studio, voice, failure_analysis}.rs
- mockforge-cli: scenario_commands.rs, voice_commands.rs

Added mockforge-foundation as a dep to mockforge-chaos, mockforge-http,
mockforge-ui, mockforge-cli.

Foundation tests: 36 → 42 passing. Workspace builds clean.
