+++ crates/mockforge-chaos/Cargo.toml
@@ -71,6 +71,7 @@ mockforge-recorder = { version = "0.3.70", path = "../mockforge-recorder" }
 
 # Core (for MockAI integration)
 mockforge-core = { version = "0.3.70", path = "../mockforge-core" }
+mockforge-foundation = { path = "../mockforge-foundation" }
 
 # PDF generation
 printpdf = "0.7"
