+++ crates/mockforge-cli/Cargo.toml
@@ -26,6 +26,7 @@ tracing = { workspace = true }
 tracing-subscriber = { workspace = true }
 tracing-opentelemetry = "0.22"
 mockforge-core = { version = "0.3.70", path = "../mockforge-core", features = ["schema"] }
+mockforge-foundation = { path = "../mockforge-foundation" }
 mockforge-import = { path = "../mockforge-import" }
 mockforge-workspace = { path = "../mockforge-workspace" }
 mockforge-http = { version = "0.3.70", path = "../mockforge-http", optional = true }
