+++ crates/mockforge-http/Cargo.toml
@@ -53,6 +53,7 @@ oauth2 = { workspace = true }
 ring = { workspace = true }
 governor = "0.8"
 mockforge-core = { version = "0.3.70", path = "../mockforge-core" }
+mockforge-foundation = { path = "../mockforge-foundation" }
 mockforge-import = { path = "../mockforge-import" }
 mockforge-proxy = { path = "../mockforge-proxy" }
 mockforge-data = { version = "0.3.70", path = "../mockforge-data" }
