+++ crates/mockforge-ui/Cargo.toml
@@ -45,6 +45,7 @@ sysinfo = { workspace = true }
 html-escape = "0.2"
 once_cell = "1.19"
 mockforge-core = { version = "0.3.70", path = "../mockforge-core" }
+mockforge-foundation = { path = "../mockforge-foundation" }
 mockforge-http = { version = "0.3.70", path = "../mockforge-http", features = ["smtp", "mqtt", "conformance"] }
 mockforge-ws = { version = "0.3.70", path = "../mockforge-ws" }
 mockforge-grpc = { version = "0.3.70", path = "../mockforge-grpc" }
